refactor and guard against invalid `tabindex`
