--- src/utils/focus/selector.ts
@@ -1,10 +1,0 @@
-export const FOCUSABLE_SELECTOR = [
-  'input:not([type=hidden]):not([disabled])',
-  'button:not([disabled])',
-  'select:not([disabled])',
-  'textarea:not([disabled])',
-  '[contenteditable=""]',
-  '[contenteditable="true"]',
-  'a[href]',
-  '[tabindex]:not([disabled])',
-].join(', ')
